Add files via upload
A study is developed in the direction of progress in predicting the winner of international cricket series. This work specifically emphasizes on forecasting the winner of T20 world cup series 2024. The objective is to predict the cricket match winning team with the respective probability percentage by using historical dataset. The purpose of the study is to predict the accurate winner so that it can benefit multiple stakeholders such as cricket team members and managers in improving their performance by analyzing important features that impact the score, improvement in fan engagement, fetching stability in betting industry, etc. These findings play major role in increasing the game popularity in sports analytics. In this analysis, python programming language and traditional machine learning algorithms and deep learning are utilized to improve the performance metrics of all ML algorithms. In the methodology, CRISP-DM data mining method is used to train and test the models. The large historic dataset is gathered from official website ESPN. The test dataset is collected from ICC cricket website. The dataset is pre-processed before training it using multiple ML models. In order to generate accurate results, Feature engineering is performed on dataset. With the help of new features different ML models such as Random Forest (RF), XG Booster, Logistic Regression (LR), and Support Vector Classifier (SVC), and deep learning were trained to evaluate their performance metrics such as accuracy, precision, recall, and F1-Score, AUC-ROC. Also, to measure the accuracy of predictive models Mean Square Error (MSE), Mean Absolute Error (MAE), and Root Mean Square Error (RMSE) are calculated. Amongst all the ML models Logistic regression proved to be the most effective model with precision of 70%, predicting India to be the winning team of the series.